requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Deprecated==1.2.18
distro==1.9.0
fastapi==0.115.12
Flask==3.1.0
Flask-Limiter==3.10.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.55.8
fuzzywuzzy==0.18.0
google-auth==2.38.0
google-genai==1.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
Levenshtein==0.26.1
limits==4.0.1
logfire-python==0.0.9
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
msgpack==1.1.0
mypy==1.14.1
mypy-extensions==1.0.0
numpy==2.2.2
openai==1.61.1
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyMuPDF==1.25.5
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
python-multipart==0.0.20
pytz==2025.1
RapidFuzz==3.12.1
requests==2.32.3
rich==13.9.4
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.46.2
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.1
websockets==15.0.1
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.17.2
WTForms==3.2.1
